+++ src/main/java/com/project/facegram/dto/CommentDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class CommentDto {
 
     private int id; // 댓글 일련번호
-    private Long postId; // 댓글이 속한 게시글의 고유 번호
+    private int postId; // 댓글이 속한 게시글의 고유 번호
     private String writer; // 댓글 작성자의 아이디
     private String content; // 댓글 내용
     private LocalDateTime date; // 댓글 작성 날짜
